Refactoring code
Movde new functions from directed_joint_degree_seq.py to joint_degree_seq.py and test_joint_degree_seq.py
